+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Extensive internal unit tests and continuous integration checks.
 * Add support for plugin.
 * Add basic shapes plugin.
+* Disable screenlocker when using apps in fullscreen mode.
 
 ### Changed
 * Freehand strokes now use Bezier curves and interpolation to make drawing feel significantly smoother.